Change in structure of PatientListData
